+++ CHANGELOG.md
@@ -19,6 +19,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Associate access tokens with OAuth applications.
 - Changed default access token lifetime to 30 days.
 - Verify public keys, identity proofs and payment options when reading profiles from database.
+- Write warning to log if reaction is followers-only.
 
 ### Fixed